Move KeyMap and icon import to utils
