+++ src/utils/icons.js
@@ -1,0 +1,36 @@
+import fontawesome from '@fortawesome/fontawesome'
+import faCheck from '@fortawesome/fontawesome-free-solid/faCheck'
+import faExclamation from '@fortawesome/fontawesome-free-solid/faExclamation'
+import faPencil from '@fortawesome/fontawesome-free-solid/faPencilAlt'
+import facolumns from '@fortawesome/fontawesome-free-solid/faColumns'
+import faDownload from '@fortawesome/fontawesome-free-solid/faCloudDownloadAlt'
+import faTrash from '@fortawesome/fontawesome-free-solid/faTrash'
+import faUndo from '@fortawesome/fontawesome-free-solid/faUndo'
+import faSearch from '@fortawesome/fontawesome-free-solid/faSearch'
+import faCheckSquare from '@fortawesome/fontawesome-free-solid/faCheckSquare'
+import faSquare from '@fortawesome/fontawesome-free-solid/faSquare'
+import faFilter from '@fortawesome/fontawesome-free-solid/faFilter'
+import faPlusSquare from '@fortawesome/fontawesome-free-solid/faPlusSquare'
+import faSave from '@fortawesome/fontawesome-free-solid/faSave'
+import faTimes from '@fortawesome/fontawesome-free-solid/faTimes'
+import faPlus from '@fortawesome/fontawesome-free-solid/faPlus'
+import faExpand from '@fortawesome/fontawesome-free-solid/faExpand'
+
+fontawesome.library.add(
+  facolumns,
+  faCheck,
+  faPencil,
+  faExclamation,
+  faTrash,
+  faDownload,
+  faUndo,
+  faSearch,
+  faSquare,
+  faCheckSquare,
+  faFilter,
+  faPlusSquare,
+  faSave,
+  faTimes,
+  faPlus,
+  faExpand
+)
